feat(ui): Add a nudge for users who have not set their security context. Closes #4233  (#4255)
